+++ pdfconduit/settings/__init__.py
@@ -1,0 +1,5 @@
+from pdfconduit.settings.compression import Compression, ImageQualityRange
+from pdfconduit.settings.encryption import Algorithms, Encryption
+
+
+__all__ = [Compression, ImageQualityRange, Encryption, Algorithms]
