+++ pdfconduit/settings/compression.py
